Added locks.
It worked somewhat for the evil sudoku, but we need more. We should be
able to eliminate '3's from the top row in box-index-2 thanks to the
"lock" of 3s in box-index-0.

I think I will call this phenomenon a "dagger" - just because it sounds
cool, like "locks and daggers" idk. I just need a name for it. The
phenomenon is when a number in a box can only be placed in a certain
orientation and thus clears pencilmarks